udpated the show route and added the express urlencoded to main index.js
