+++ index.js
@@ -1,6 +1,6 @@
 const express= require('express')
 const app = express() 
-const cors = require('cors')
+
 
 require('dotenv').config() 
 const { MongoClient } = require("mongodb");
@@ -11,11 +11,13 @@ app.set("views", __dirname + "/views");
 app.set('view engine','jsx') 
 app.engine('jsx',require('express-react-views').createEngine())
 app.use(express.static('public'))
-app.use('/topics', require('./controllers/topic_controllers'))
-app.use(methodOverride('_method'))
-app.use(cors())
+app.use(express.urlencoded({extended:true})); 
 app.use(express.json())
-app.use(express.urlencoded({ extended: false })); 
+
+app.use(methodOverride('_method'))
+
+
+app.use('/topics', require('./controllers/topic_controllers'))
 
 app.get('/',(req,res)=>{
     res.render('home')
