+++ public/style.css
@@ -9,7 +9,7 @@ h1 {
  
   h2 {
     text-align: center;
-    color: #ff6c2c;
+    color: #696969;
     margin-top: 15px;
   }
   h5 {
