+++ views/resources/edit.jsx
@@ -1,4 +1,4 @@
-import React,{useState} from 'react' 
+import React from 'react' 
 const Def = require('../default')
