+++ views/resources/tempcode.jsx
@@ -1,35 +1,35 @@
-import React,{useState} from'react' 
 
-function TestForm =new FormData
-const [input,setInput]=useState({
-    resources_name:'',
-    resources_type:'',
-    resources_link:'',
-    resources_id: '',
-    resources_topic_id:''
-})
-
-function handleChange(e){
-    const {name,value}=e.target;
-
-    setInput(prevInput =>{
-        return{
-            ...prevInput,
-            [name]:value
+    const[values, setValues] = useState({ 
+        subject_name:'',
+        topic_name:'',
+        resources_id: '',
+        resouces_name: '', 
+        resources_link:'', 
+        resources_type: '',
+        resources_topic_id:'' 
+      });
+      
+      const saveFormData = async () => {
+        const response = await fetch('/resources/new', {
+          method: 'POST',
+          body: JSON.stringify(values)
+        });
+        if (response.status !== 200) {
+          throw new Error(`Request failed: ${response.status}`); 
         }
-    })
-}
-function handleSubmit(e){
-    e.preventDefault;
-    const url = "`/topics/${data.subjectName}/${data.topicName}/${data.topicId}/resources/${data.resource.resources_id}?_method=PUT`";
-    const formData = new formData(e.target);
-    const data2 = {};
-    formData.forEach((value, key) => (data[key] = value));
-
-    // Submit the data.
-    //const request = new XMLHttpRequest();
-   /* request.open("POST", url);
-    request.send(formData);*/
-}
-
-module.exports = editResource
+      }
+      const handleChange= (e)=>{
+        this.setState({name: e.target.value });
+      };
+      const onSubmit = async (e) => {
+        e.preventDefault(); // Prevent default submission
+        try {
+          await saveFormData();
+          alert('Your registration was successfully submitted!');
+          setValues({
+            topic_name:'',resources_id:'', resources_name: '', resources_link: '', resources_type: '', resources_topic_id:'' 
+          });
+        } catch (e) {
+          alert(`Registration failed! ${e.message}`);
+        }
+        }
